Create models for the cricpro app
This commit introduces the following models for the cricpro app:
    * Team
    * Player
    * PlayerHistory
    * Matches
    * Points

The models are as per specification given with the exception of
PlayerHistory. This data being variable with possibility of adding
multiple metadata fields for players over time and sometimes missing
some or more of the fields, a generic key, value model has been kept for
the same.

The team is likely the central piece in this app, with multiple teams
having multiple players, mutiple matches, multiple points-pairs and vice
versa.